Refactor Home component to use HomePage; remove unused user mock data and related components; update ClearButton export name; enhance DropdownMenu styles; adjust Header component layout and avatar size; update IComponents types for better type safety.
