+++ src/components/Atoms/ClearButton/index.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react';
 import styles from './index.module.scss';
 import LucideIcon from '@/components/Atoms/LucideIcon';
 
-export default function Button({
+export default function ClearButton({
     children,
     width,
     height,
