+++ src/components/Molecules/DropdownMenu/index.module.scss
@@ -117,6 +117,22 @@
   box-shadow: 0px 10px 38px -10px rgba(22, 23, 24, 0.35), 0px 10px 20px -15px rgba(22, 23, 24, 0.2);
   animation-duration: 400ms;
   animation-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
+  margin-top: 10px;
+  position: relative;
+  border: 1px solid rgba(22, 23, 24, 0.1);
+
+  &::before {
+    content: '';
+    position: absolute;
+    top: -8px;
+    right: 44px;
+    width: 14px;
+    height: 14px;
+    background-color: white;
+    transform: rotate(45deg);
+    border-left: 1px solid rgba(22, 23, 24, 0.1);
+    border-top: 1px solid rgba(22, 23, 24, 0.1);
+  }
 }
 
 .dropdown_label {
