+++ src/components/Organisms/Header/index.tsx
@@ -13,10 +13,10 @@ import {
 
 export default function Header() {
     return (
-        <header className="bg-grey p-3 px-8 w-full rounded-xl">
+        <header className="bg-grey p-2 px-8 w-full rounded-md">
             <div className="flex justify-between items-center">
                 <Link href={ROUTES.PUBLIC.HOME}>
-                    <Image src="https://res.cloudinary.com/dodtzdovx/image/upload/v1744187841/photogo_black_otpabv.svg" alt="logo" width={70} height={70} style={{ width: 'auto', height: 'auto' }} priority />
+                    <Image src="https://res.cloudinary.com/dodtzdovx/image/upload/v1744187841/photogo_black_otpabv.svg" alt="logo" width={60} height={60} style={{ width: 'auto', height: 'auto' }} priority />
                 </Link>
 
                 <div className="flex gap-12 font-medium text-md">
@@ -32,7 +32,7 @@ export default function Header() {
                             <Avatar
                                 src="https://thanhnien.mediacdn.vn/Uploaded/haoph/2021_10_21/jack-va-thien-an-5805.jpeg"
                                 alt="User avatar"
-                                size={50}
+                                size={40}
                                 className="cursor-pointer"
                             />
                         </DropdownMenuTrigger>
