--- src/pages/HomePage/index.tsx
@@ -1,9 +1,0 @@
-'use client'
-
-const HomePage = () => {
-    return (
-        <div>HomePage</div>
-    )
-}
-
-export default HomePage
